Fix handling of closed mysql socket.
mysql_recv sends mysql_conn a `{mysql_recv, self(), closed, Error}`
message and exits normally when it notices that the mysql connection
has been closed.  Unfortunately, mysql_conn handles these "mysql_recv
closed" messages in several places but only itself exits if the
"mysql_recv closed" message is received in mysql_conn:loop (where it
is handled in the "Unknown" clause).  If the "mysql_recv closed"
message is received in another location, the fact that the mysql_recv
process has exited is lost.

Furthermore, although mysql_conn is linked to mysql_recv, when
mysql_recv notices the closed connection it exits normally.  That
means that mysql_conn ignores the exit signal from mysql_recv.

Finally, the mysql gen_server only removes the mysql socket from the
connection pool if the associated mysql_conn process itself
exits.  (The mysql gen_server is `monitor`ing the mysql_conn process
and receives a `DOWN` message when that happens.)  That means that
although the mysql request that discovers that the connection has been
closed will return a "closed" error, the closed connection may be
returned to the connection pool to be used by the next request.  At
this point, the next next request to use that connection will notice
that it is closed in the `gen_tcp:send` call and the code path in
mysql_conn will never exit, meaning that this connection will just
error forever.

This code change fixes the problem because it causes the mysql_conn
process to also receive a `DOWN` message when the mysql_recv process
exits and this message is received in only one place in the mysql_conn
process: in the "Unknown" clause in mysql_conn:loop.  That means that
mysql_conn will always notice that the mysql_recv process has exited
and will always exit itself.

This code change also has the advantage over having mysql_recv just
exit with a non-`normal` exit status that the caller will always
receive a "closed" error instead of just having the call timeout,
which could happen if mysql_conn exited because of mysql_recv's
non-`normal` exit status before it handled the "mysql_recv closed"
message.